refactor!: all commands rewritten
test: remove old command tests
chore: remove old User model
fix: asset routes now use correct path
