--- src/commands/index.ts
@@ -1,6 +1,0 @@
-import ListCommands from "./ListCommands"
-import BeatAss from "./BeatAss"
-import Beans from "./memes/Beans"
-import Socials from "./Socials"
-
-export { ListCommands, BeatAss, Beans, Socials }
