+++ src/index.ts
@@ -1,49 +1,28 @@
-import logger, { format, transports } from "winston"
-import { enqueueNewAlert, MoneySystem, setSocket } from "./yuki"
+import {
+  Beans,
+  BeatAss,
+  enqueueNewAlert,
+  FitCheck,
+  Fox,
+  Hydrate,
+  ListCommands,
+  MoneySystem,
+  Pushups,
+  setSocket,
+  Socials,
+  Wallet,
+} from "./yuki"
 import ENV from "./env"
-import "./testing"
 import { yuki, YukiBuilder } from "@pinkilo/yukibot"
-import { Beans, BeatAss, ListCommands, Socials } from "./commands"
 import { WebSocketServer } from "ws"
 import { file } from "./util"
 import { Credentials } from "google-auth-library"
 import server from "./server"
+import logger from "./logger"
 
-logger.configure({
-  level: ENV.NODE_ENV === "test" ? "debug" : "info",
-  transports: [new transports.Console()],
-  format: format.combine(
-    format.errors({ stack: true }),
-    format.colorize({ all: true }),
-    format.timestamp(),
-    format.printf(
-      (info) =>
-        `[${info.timestamp}] YUKI ${info.level}: ${info.message} ${
-          info.err || ""
-        }`
-    )
-  ),
-})
-
-const TokenHandlers = (y: YukiBuilder) => {
-  y.tokenLoader = async () => {
-    logger.info("checking for saved tokens")
-    if (file.exists(ENV.FILE.TOKENS)) {
-      logger.debug("attempting to parse saved tokens")
-      const raw = await file.read(ENV.FILE.TOKENS)
-      let tokens: Credentials = JSON.parse(raw.toString())
-      return tokens
-    }
-    logger.info("No saved tokens")
-    return undefined
-  }
-  y.onAuthUpdate(async (tokens) => {
-    await file.write(ENV.FILE.TOKENS, JSON.stringify(tokens))
-  })
-}
-
-const Commands = async (y: YukiBuilder) =>
-  Promise.all([BeatAss(y), ListCommands(y), Beans(y), Socials(y)])
+main()
+  .then()
+  .catch((e) => logger.error("crashed", { e }))
 
 async function main() {
   logger.info(`Running in ${ENV.NODE_ENV}`)
@@ -52,9 +31,9 @@ async function main() {
   await MoneySystem.walletCache.load(ENV.FILE.CACHE.BANK)
 
   const bot = await yuki(async (y) => {
-    y.logLevel = "http"
     y.yukiConfig.name = "Yuki"
     y.yukiConfig.prefix = /^[>!].*$/
+    y.loggerOverride = logger
     y.googleConfig = {
       clientId: ENV.GOOGLE.G_CLIENT_ID,
       clientSecret: ENV.GOOGLE.G_CLIENT_SECRET,
@@ -81,6 +60,37 @@ async function main() {
   setSocket(new WebSocketServer({ server: svr, path: "/fox" }))
 }
 
-main()
-  .then()
-  .catch((e) => logger.error("crashed", { e }))
+const TokenHandlers = (y: YukiBuilder) => {
+  y.tokenLoader = async () => {
+    logger.info("checking for saved tokens")
+    if (file.exists(ENV.FILE.TOKENS)) {
+      logger.info("attempting to parse saved tokens")
+      const raw = await file.read(ENV.FILE.TOKENS)
+      let tokens: Credentials = JSON.parse(raw.toString())
+      return tokens
+    }
+    logger.info("no saved tokens")
+    return undefined
+  }
+  y.onAuthUpdate(async (tokens) => {
+    logger.info("writing tokens to file")
+    await file.write(ENV.FILE.TOKENS, JSON.stringify(tokens))
+  })
+}
+
+const Commands = async (y: YukiBuilder) =>
+  Promise.all([
+    BeatAss(y),
+    ListCommands(y),
+    Beans(y),
+    Socials(y),
+    Fox.Attack(y),
+    Fox.Feed(y),
+    Fox.Dance(y),
+    FitCheck(y),
+    Pushups(y),
+    Hydrate(y),
+    Wallet.Ranking(y),
+    Wallet.View(y),
+    Wallet.Leaderboard(y),
+  ])
