--- src/models/index.ts
@@ -1,3 +1,0 @@
-import User from "./User"
-
-export { User }
