+++ src/yuki/_commands/ListCommands.ts
@@ -1,0 +1,8 @@
+import Command from "./Command"
+import yt from "../../youtube"
+
+export default new Command("commands", ["cmds"], 0, 0, 360, async () => {
+  await yt.chat.sendMessage(
+    "Find a list of _commands here https://tinyurl.com/mt8c5v47"
+  )
+})
