+++ src/yuki/commands/BeatAss.ts
@@ -1,40 +1,40 @@
-import yt from "../../youtube"
-import Command from "./Command"
+import { YukiBuilder } from "@pinkilo/yukibot"
+import { randomUser } from "../../util/yukiExt"
+import supercommand from "./supercommand"
 import logger from "winston"
+import MS from "../MoneySystem"
 
-/**
- * Randomly selects another chat member to "fight"
- * Rolls odds (for now 50:50)
- * Announces winner in chat
- *
- * TODO Additional text version
- */
-export default new Command(
-  "beatass",
-  ["pickfight"],
-  10,
-  60,
-  0,
-  async ({ authorDetails: { displayName, channelId } }, _, cost, _this) => {
-    const rUser = yt.users.getRandomUser([channelId])
-    const { id: tid, name } = rUser
-    logger.debug("running beatass", { target: name, displayName })
-    const succeeds = Math.random() > 0.55
-    const successPayout = cost * 2
-    const defensePayout = cost
-    const failed = await yt.chat.sendMessage(
-      succeeds
-        ? `${displayName} beat ${name} 's ass (+${successPayout})`
-        : `${displayName} tried to beat ${name} 's ass but failed and got 
+export default (y: YukiBuilder) =>
+  y.command(async (c) => {
+    c.name = "beatass"
+    c.alias = ["pickfight"]
+    c.rateLimit.individual = 60
+    supercommand(
+      c,
+      10,
+      async ({ authorDetails: { displayName, channelId } }, _, cost) => {
+        const rUser = randomUser(y, channelId)
+        const { id: tid, name } = rUser
+        logger.debug("running beatass", { target: name, displayName })
+        const winsFight = Math.random() > 0.55
+        const successPayout = cost * 2
+        const defensePayout = cost
+        const { success } = await y.sendMessage(
+          winsFight
+            ? `${displayName} beat ${name} 's ass (+${successPayout})`
+            : `${displayName} tried to beat ${name} 's ass but failed and got 
           the shit smacked outta them (+${defensePayout} to the defender)`
+        )
+        if (!success) {
+          logger.error("failed to send message")
+          return
+        }
+        await MS.transactionBatch([
+          [
+            winsFight ? channelId : tid,
+            winsFight ? successPayout : defensePayout,
+          ],
+        ])
+      }
     )
-    if (failed) {
-      logger.error("failed to send message")
-      return undefined
-    }
-    return {
-      uids: [succeeds ? channelId : tid],
-      amount: succeeds ? successPayout : defensePayout,
-    }
-  }
-)
+  })
