+++ src/yuki/commands/ListCommands.ts
@@ -1,8 +1,12 @@
-import Command from "./Command"
-import yt from "../../youtube"
+import { YukiBuilder } from "@pinkilo/yukibot"
 
-export default new Command("commands", ["cmds"], 0, 0, 360, async () => {
-  await yt.chat.sendMessage(
-    "Find a list of commands here https://tinyurl.com/mt8c5v47"
-  )
-})
+export default async (y: YukiBuilder) =>
+  y.command(async (c) => {
+    c.name = "commands"
+    c.alias = ["cmds", "cmd"]
+    c.rateLimit.global = 360
+    c.invoke = async () =>
+      await y.sendMessage(
+        "Find a list of _commands here https://tinyurl.com/mt8c5v47"
+      )
+  })
