+++ src/yuki/commands/Wallet.ts
@@ -1,69 +1,54 @@
-import Command from "./Command"
-import yt from "../../youtube"
+import { YukiBuilder } from "@pinkilo/yukibot"
 import MS from "../MoneySystem"
-import logger from "winston"
 import { enqueueNewAlert } from "../alerts"
 
-export const Wallet = new Command(
-  "wallet",
-  ["bank"],
-  0,
-  120,
-  0,
-  async ({ authorDetails }, tokens) => {
-    let msg: string
-    switch (tokens.params[0]) {
-      case undefined:
-      default:
-        const wallet = MS.walletCache.get(authorDetails.channelId)
-        msg = `${authorDetails.displayName} has ${wallet} ${MS.name}s`
-        break
+export const View = (y: YukiBuilder) =>
+  y.command((c) => {
+    c.name = "wallet"
+    c.alias = ["bank"]
+    c.rateLimit.individual = 60 * 2
+    c.invoke = async ({ authorDetails }) => {
+      const wallet = MS.walletCache.get(authorDetails.channelId)
+      let msg = `${authorDetails.displayName} has ${wallet} ${MS.name}s`
+      await y.sendMessage(msg)
     }
-    const failed = await yt.chat.sendMessage(msg)
-    if (failed) logger.error("failed to send message")
-  }
-)
+  })
 
-/**
- * If RANK send ranking
- * else if WEALTHGAP send number of people below you in rank
- */
-export const Ranking = new Command(
-  "rank",
-  ["wealthgap"],
-  0,
-  120,
-  0,
-  async ({ authorDetails: { channelId, displayName } }, { command }) => {
-    const wallet = MS.walletCache.get(channelId)
-    const lb = await MS.getLeaderboard()
-    const rank = lb.findIndex(([uid]) => uid === channelId)
-    let msg =
-      command == "rank"
-        ? `#${rank + 1}: ${displayName} | ${wallet} ${MS.name}`
-        : `${lb.length - rank - 1} citizen(s) are poorer than ${displayName}`
-    const failed = await yt.chat.sendMessage(msg)
-    if (failed) logger.error("failed to send message")
-  }
-)
+export const Ranking = (y: YukiBuilder) =>
+  y.command((c) => {
+    c.name = "rank"
+    c.alias = ["wealthgap"]
+    c.rateLimit.individual = 60 * 2
+    c.invoke = async (
+      { authorDetails: { channelId, displayName } },
+      { command }
+    ) => {
+      const wallet = MS.walletCache.get(channelId)
+      const lb = await MS.getLeaderboard()
+      const rank = lb.findIndex(([uid]) => uid === channelId)
+      let msg =
+        command == "rank"
+          ? `#${rank + 1}: ${displayName} | ${wallet} ${MS.name}`
+          : `${lb.length - rank - 1} citizen(s) are poorer than ${displayName}`
+      await y.sendMessage(msg)
+    }
+  })
 
 let leaderboardDisplayDuration: number = 0
 
-export const popLeaderboardDisplayTimer = (): number | null => {
+export const popLeaderboardDisplayTimer = (): number => {
   const out = leaderboardDisplayDuration
   leaderboardDisplayDuration = 0
   return out
 }
 
-/** @command leaderboard [me] set leaderboardDisplay to true for N seconds */
-export const Leaderboard = new Command(
-  "leaderboard",
-  ["forbes"],
-  10,
-  0,
-  60 * 3,
-  async ({ authorDetails: { channelId, displayName } }) => {
-    leaderboardDisplayDuration = 30
-    await enqueueNewAlert("Leaderboard Display", displayName, channelId)
-  }
-)
+export const Leaderboard = (y: YukiBuilder) =>
+  y.command((c) => {
+    c.name = "leaderboard"
+    c.alias = ["forbes"]
+    c.rateLimit.global = 60 * 3
+    c.invoke = async ({ authorDetails: { channelId, displayName } }) => {
+      leaderboardDisplayDuration = 30
+      await enqueueNewAlert("Leaderboard Display", displayName, channelId)
+    }
+  })
