+++ src/yuki/commands/supercommand.ts
@@ -1,7 +1,7 @@
 import { Command, CommandBuilder, TokenBin } from "@pinkilo/yukibot"
-import { ChatMessage } from "../types/google"
+import { ChatMessage } from "../../types/google"
 import { youtube_v3 } from "googleapis"
-import MS from "../yuki/MoneySystem"
+import MS from "../MoneySystem"
 import logger from "winston"
 import Schema$LiveChatMessage = youtube_v3.Schema$LiveChatMessage
